requirements.txt
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
fake-useragent==1.2.1
idna==3.4
lxml==4.9.3
psycopg2-binary==2.9.7
python-dotenv==1.0.0
requests==2.31.0
soupsieve==2.5
urllib3==2.0.4
pytest==7.4.2